+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 ## 0.1.6.2 (In progress)
  - Timber updates
- - HTML editor
+ - Ability to toggle HTML editor
+ - DarkLeather theme (init)
 
 ## 0.1.6.1
  - Css bugfix (affects Timber only)